+++ crypto/err/openssl.ec
@@ -39,7 +39,7 @@ L SM2           include/crypto/sm2.h            crypto/sm2/sm2_err.c
 L OSSL_STORE    include/openssl/store.h         crypto/store/store_err.c
 L ESS           include/openssl/ess.h           crypto/ess/ess_err.c
 L PROP          include/internal/property.h     crypto/property/property_err.c
-L PROV          providers/common/include/internal/providercommon.h providers/common/provider_err.c
+L PROV          providers/common/include/prov/providercommon.h providers/common/provider_err.c
 
 # additional header files to be scanned for function names
 L NONE          include/openssl/x509_vfy.h      NONE
