+++ providers/common/ciphers/cipher_gcm.c
@@ -11,9 +11,9 @@
 
 #include "prov/ciphercommon.h"
 #include "prov/cipher_gcm.h"
-#include "internal/providercommonerr.h"
+#include "prov/providercommonerr.h"
 #include "crypto/rand.h"
-#include "internal/provider_ctx.h"
+#include "prov/provider_ctx.h"
 
 static int gcm_tls_init(PROV_GCM_CTX *dat, unsigned char *aad, size_t aad_len);
 static int gcm_tls_iv_set_fixed(PROV_GCM_CTX *ctx, unsigned char *iv,
