+++ providers/fips/build.info
@@ -1,3 +1,3 @@
 
 SOURCE[../fips]=fipsprov.c selftest.c
-INCLUDE[../fips]=../common/include
+INCLUDE[../fips]=../implementations/include ../common/include
