+++ providers/implementations/ciphers/cipher_aes_hw.c
@@ -8,7 +8,7 @@
  */
 
 #include "cipher_aes.h"
-#include "internal/providercommonerr.h"
+#include "prov/providercommonerr.h"
 
 static int cipher_hw_aes_initkey(PROV_CIPHER_CTX *dat,
                                  const unsigned char *key, size_t keylen)
