+++ providers/implementations/ciphers/cipher_des.c
@@ -11,7 +11,7 @@
 #include "cipher_des.h"
 #include "crypto/rand.h"
 #include "prov/implementations.h"
-#include "internal/providercommonerr.h"
+#include "prov/providercommonerr.h"
 
 /* TODO(3.0) Figure out what flags need to be here */
 #define DES_FLAGS (EVP_CIPH_RAND_KEY)
