+++ providers/implementations/ciphers/cipher_tdes.c
@@ -11,7 +11,7 @@
 #include "cipher_tdes.h"
 #include "crypto/rand.h"
 #include "prov/implementations.h"
-#include "internal/providercommonerr.h"
+#include "prov/providercommonerr.h"
 
 void *tdes_newctx(void *provctx, int mode, size_t kbits, size_t blkbits,
                   size_t ivbits, uint64_t flags, const PROV_CIPHER_HW *hw)
