+++ providers/implementations/digests/mdc2_prov.c
@@ -14,7 +14,7 @@
 #include <openssl/err.h>
 #include "prov/digestcommon.h"
 #include "prov/implementations.h"
-#include "internal/providercommonerr.h"
+#include "prov/providercommonerr.h"
 
 static OSSL_OP_digest_set_ctx_params_fn mdc2_set_ctx_params;
 static OSSL_OP_digest_settable_ctx_params_fn mdc2_settable_ctx_params;
