+++ providers/implementations/kdfs/kbkdf.c
@@ -38,9 +38,9 @@
 #include "crypto/evp.h"
 #include "internal/numbers.h"
 #include "prov/implementations.h"
-#include "internal/provider_ctx.h"
-#include "internal/provider_util.h"
-#include "internal/providercommonerr.h"
+#include "prov/provider_ctx.h"
+#include "prov/provider_util.h"
+#include "prov/providercommonerr.h"
 
 #include "e_os.h"
