+++ providers/implementations/kdfs/x942kdf.c
@@ -25,10 +25,10 @@
 # include "internal/cryptlib.h"
 # include "internal/numbers.h"
 # include "crypto/evp.h"
-# include "internal/provider_ctx.h"
-# include "internal/providercommonerr.h"
+# include "prov/provider_ctx.h"
+# include "prov/providercommonerr.h"
 # include "prov/implementations.h"
-# include "internal/provider_util.h"
+# include "prov/provider_util.h"
 
 # define X942KDF_MAX_INLEN (1 << 30)
