+++ providers/implementations/macs/gmac_prov.c
@@ -15,10 +15,10 @@
 #include <openssl/evp.h>
 #include <openssl/err.h>
 
-#include "internal/providercommonerr.h"
+#include "prov/providercommonerr.h"
 #include "prov/implementations.h"
-#include "internal/provider_ctx.h"
-#include "internal/provider_util.h"
+#include "prov/provider_ctx.h"
+#include "prov/provider_util.h"
 
 /*
  * Forward declaration of everything implemented here.  This is not strictly
